refactor: incrementata modularità del codice
