+++ src/handlers/load_file.py
@@ -1,0 +1,17 @@
+from telegram.ext import CallbackContext
+from src.data.costanti import LIVELLO, CATEGORIA
+import json
+
+def load_file(context: CallbackContext) -> dict:
+    match context.user_data[LIVELLO]:
+        case "1: Facile":
+            path = 'facile.json'
+        case "2: Intermedio":
+            path = 'intermedio.json'
+        case "3: Difficile":
+            path = 'difficile.json'
+
+    with open(f'src/data/{path}', 'r', encoding="utf-8") as f:
+        data = json.load(f)[f"{context.user_data[CATEGORIA]}"]
+
+    return data
